# server/service/redfish/openapi.yaml
openapi: 3.1.0
info:
  title: NanoKVM BMC Redfish API
  version: 1.0.0
  summary: Out-of-band management for Raspberry Pi 5 via the Redfish standard.
  description: |
    This is the Redfish (DMTF DSP0266) surface the NanoKVM BMC exposes for
    out-of-band management of a managed Raspberry Pi 5.

    The schemas here describe the **subset** of standard Redfish this BMC
    implements, plus the BMC-specific BIOS layer that maps RPi 5 bootloader
    EEPROM configuration (the `bootconf.txt` `[all]` section) onto Redfish
    `Bios.Attributes`.

    For the full Redfish data-model reference, see
    https://redfish.dmtf.org/schemas/v1/ — every `@odata.type` returned by
    this BMC is a real DMTF schema name.

    ## Authentication

    Public endpoints (no auth):
      - `GET  /redfish`
      - `GET  /redfish/v1` (ServiceRoot)
      - `GET  /redfish/v1/SessionService`
      - `POST /redfish/v1/SessionService/Sessions` (create a session)

    All other endpoints require either:
      - an `X-Auth-Token` header carrying a session token, or
      - the JWT cookie issued by the web UI's `/api/auth/login` flow.

    Tinkerbell / `bmclib` callers should create a session at
    `/redfish/v1/SessionService/Sessions` and re-use the returned token.

  contact:
    name: NanoKVM
    url: https://github.com/pi-bmc/nanokvm-app
servers:
  - url: /
    description: The BMC itself.
...
security:
  - XAuthToken: []
  - BMCCookie: []
paths:
...
  /redfish/v1/Systems/1:
    get:
      tags: [Systems]
      summary: The managed Raspberry Pi 5 — power state, boot override, inventory.
      responses:
        "200":
          description: ComputerSystem document.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ComputerSystem" }
    patch:
      tags: [Systems]
      summary: Change boot override (sets U-Boot env on the FAT).
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                Boot:
                  type: object
                  properties:
                    BootSourceOverrideTarget:
                      type: string
                      enum: [None, Pxe, Hdd, Cd, BiosSetup, UefiHttp]
                    BootSourceOverrideEnabled:
                      type: string
                      enum: [Disabled, Once, Continuous]
              example:
                Boot:
                  BootSourceOverrideTarget: Pxe
                  BootSourceOverrideEnabled: Once
      responses:
        "200":
          description: New ComputerSystem state.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ComputerSystem" }
        "400": { $ref: "#/components/responses/BadRequest" }
...
  /redfish/v1/Managers/1/VirtualMedia/CD/Actions/VirtualMedia.InsertMedia:
    post:
      tags: [Managers]
      summary: Insert an ISO image as a USB CD/disk to the managed host.
      description: |
        Two transfer methods are supported:

        * **Stream** (default) — `application/json` body names an HTTP(S)
          URL; the BMC pulls the image.
        * **Upload** — `multipart/form-data` push from the client. Include
          the binary as a file part named `Image` (or `file`), plus an
          optional `InsertMediaRequestBody` JSON part to override the
          staged filename.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [Image]
              properties:
                Image:
                  type: string
                  description: HTTP(S) URL to the ISO.
                TransferMethod:
                  type: string
                  enum: [Stream]
                  default: Stream
                Inserted: { type: boolean }
                WriteProtected: { type: boolean }
              example:
                {
                  Image: "https://releases.ubuntu.com/22.04/ubuntu-22.04-live-server-arm64.iso",
                }
          multipart/form-data:
            schema:
              type: object
              required: [Image]
              properties:
                Image:
                  type: string
                  format: binary
                  description: The image file body (also accepted as `file`).
                InsertMediaRequestBody:
                  type: string
                  description: |
                    Optional JSON metadata part. Same shape as the
                    `application/json` body; useful to override the staged
                    filename via `Image`.
                  example: '{"TransferMethod":"Upload","Image":"ubuntu.iso"}'
      responses:
        "200": { description: Media inserted. }
        "400": { $ref: "#/components/responses/BadRequest" }
        "500": { $ref: "#/components/responses/ServerError" }
...
components:
  securitySchemes:
    XAuthToken:
      type: apiKey
      in: header
      name: X-Auth-Token
      description: Session token issued by `POST /redfish/v1/SessionService/Sessions`.
    BMCCookie:
      type: apiKey
      in: cookie
      name: nano-kvm-token
      description: JWT cookie issued by the web UI's login flow.
...
  responses:
...
    BadRequest:
      description: Request couldn't be parsed or contained an invalid value.
      content:
        application/json:
          schema: { $ref: "#/components/schemas/RedfishError" }
    ServerError:
      description: BMC-side failure (FS error, hardware control failure, etc.).
      content:
        application/json:
          schema: { $ref: "#/components/schemas/RedfishError" }
  schemas:
...
    ComputerSystem:
      type: object
      properties:
        "@odata.type":
          { type: string, example: "#ComputerSystem.v1_13_0.ComputerSystem" }
        "@odata.id": { type: string, example: "/redfish/v1/Systems/1" }
        Id: { type: string, example: "1" }
        Name: { type: string }
        SystemType: { type: string, example: Physical }
        PowerState: { type: string, enum: ["On", "Off"] }
        Model: { type: string }
        Manufacturer: { type: string }
        SerialNumber: { type: string }
        MACAddress: { type: string }
        FirmwareVersion: { type: string }
        ProcessorSummary:
          type: object
          properties:
            Model: { type: string }
        Boot:
          type: object
          properties:
            BootSourceOverrideTarget:
              { type: string, enum: [None, Pxe, Hdd, Cd, BiosSetup, UefiHttp] }
            BootSourceOverrideEnabled:
              { type: string, enum: [Disabled, Once, Continuous] }
            "BootSourceOverrideTarget@Redfish.AllowableValues":
              type: array
              items: { type: string }
            "BootSourceOverrideEnabled@Redfish.AllowableValues":
              type: array
              items: { type: string }
            # Present when the UEFI variable store (I2C EEPROM) is available:
            # the EFI boot manager BootOrder as Boot#### references.
            BootOrder:
              type: array
              items: { type: string }
        Actions:
          type: object
          properties:
            "#ComputerSystem.Reset":
              type: object
              properties:
                target: { type: string }
                "ResetType@Redfish.AllowableValues":
                  type: array
                  items: { type: string }
        Bios:
          type: object
          properties:
            "@odata.id": { type: string, example: "/redfish/v1/Systems/1/Bios" }
...
    RedfishError:
      type: object
      description: |
        Standard Redfish error envelope (DSP0266 §6.5.6). The `code` field
        is a DMTF-defined Message ID; `message` is the human-readable text.
      properties:
        error:
          type: object
          properties:
            code: { type: string, example: "Base.1.13.GeneralError" }
            message: { type: string }
            "@Message.ExtendedInfo":
              type: array
              items:
                type: object
                properties:
                  MessageId: { type: string }
                  Message: { type: string }
                  Severity: { type: string, enum: [OK, Warning, Critical] }